requirements.txt
fastapi
uvicorn[standard]
httpx
pytest
python-dotenv
pydantic-settings
mysql-connector-python